+++ cli/src/commands/init/index.test.ts
@@ -58,8 +58,16 @@ const fetchImplementation = (url: string, request: any) => {
       ok: true,
       json: async () => ({ meta: { cursor: '', more: false }, logs: [] })
     };
+  } else if (url === `https://test-1234.${REGION}.xata.sh/dbs/db1` && request.method === 'GET') {
+    return {
+      ok: true,
+      json: async () => {
+        return {
+          branches: [{ name: 'main', id: 'main' }]
+        };
+      }
+    };
   }
-
   throw new Error(`Unexpected fetch request: ${url} ${request.method}`);
 };
 
@@ -214,7 +222,9 @@ describe('xata init', () => {
     expect(promptsMock.mock.calls.flat().find((p) => p.name === 'packageManagerName')).toBeFalsy();
     expect(warn.mock.calls.flat()).toEqual(
       expect.arrayContaining([
-        expect.stringContaining('No package.json found. Please run one of: pnpm init, yarn init, npm init. Then rerun')
+        expect.stringContaining(
+          'No package.json found. Please run one of: pnpm init, yarn init, npm init, bun init. Then rerun'
+        )
       ])
     );