+++ cli/src/commands/pull/index.ts
@@ -21,6 +21,10 @@ export default class Pull extends BaseCommand<typeof Pull> {
       char: 'f',
       description: 'Overwrite local migrations',
       default: false
+    }),
+    'skip-code-generation': Flags.boolean({
+      description: 'Skip code generation',
+      default: false
     })
   };
 
@@ -64,7 +68,7 @@ export default class Pull extends BaseCommand<typeof Pull> {
 
     this.log(`Successfully pulled ${newMigrations.length} migrations from ${branch} branch`);
 
-    if (this.projectConfig?.codegen) {
+    if (this.projectConfig?.codegen && !flags['skip-code-generation']) {
       this.log(`Running codegen...`);
       await Codegen.run(['--branch', branch]);
     }